Day11: added functionality for part 2 - somewhere detection seems to be inacurate on higher values -> some items do go to the wrong monkey

+++ aoc11/src/main/kotlin/de/havox_design/aoc2022/day11/Item.kt
@@ -1,3 +1,3 @@
 package de.havox_design.aoc2022.day11
 
-data class Item(var worryLevel: Int)
+data class Item(var worryLevel: Long)
